Externalizing TIE DXL Connector class
