+++ .env
@@ -18,4 +18,4 @@ LASTLINE_USER = 'email address'
 LASTLINE_PW = 'password'
 
 # TIE DXL Connector settings
-DXL_CONNECTOR_DXL_CONFIG = 'dxlclient.config'
+DXL_CONNECTOR_CLIENT_CONFIG_PATH = 'dxlclient.config'
